+++ entity/src/epilot/models/shared/getrelationsrespwithpagination.py
@@ -2,10 +2,11 @@
 
 from __future__ import annotations
 import dataclasses
+from ..shared import getrelationsresp as shared_getrelationsresp
 from ..shared import relationitem as shared_relationitem
 from dataclasses_json import Undefined, dataclass_json
 from epilot import utils
-from typing import Any, Optional, Union
+from typing import Optional, Union
 
 
 @dataclass_json(undefined=Undefined.EXCLUDE)